refactor(token): add missing license, remove comments
