+++ hedera-node/src/main/java/com/hedera/services/txns/token/TokenTransferTransitionLogic.java
@@ -33,7 +33,7 @@ import org.apache.logging.log4j.Logger;
 import java.util.function.Function;
 import java.util.function.Predicate;
 
-import static com.hedera.services.txns.validation.TokenChecks.checkTokenTransfers;
+import static com.hedera.services.txns.validation.TokenListChecks.checkTokenTransfers;
 import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.FAIL_INVALID;
 import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.OK;
 import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.SUCCESS;
