+++ hedera-node/src/main/java/com/hedera/services/txns/token/TokenUpdateTransitionLogic.java
@@ -39,8 +39,8 @@ import java.util.function.Function;
 import java.util.function.Predicate;
 
 import static com.hedera.services.tokens.TokenStore.MISSING_TOKEN;
-import static com.hedera.services.txns.validation.TokenChecks.checkKey;
-import static com.hedera.services.txns.validation.TokenChecks.checkKeys;
+import static com.hedera.services.txns.validation.TokenListChecks.checkKey;
+import static com.hedera.services.txns.validation.TokenListChecks.checkKeys;
 import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.FAIL_INVALID;
 import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.INVALID_FREEZE_KEY;
 import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.INVALID_TOKEN_ID;
